Move plot->grpc serializer. This will be rewritten soon with custom objs
