+++ grpc-api/src/main/java/io/deephaven/grpc_api/console/ConsoleServiceGrpcImpl.java
@@ -6,23 +6,21 @@ package io.deephaven.grpc_api.console;
 
 import com.google.rpc.Code;
 import io.deephaven.configuration.Configuration;
-import io.deephaven.engine.util.jpy.JpyInit;
-import io.deephaven.jpy.JpyConfig;
-import io.deephaven.plot.FigureWidget;
 import io.deephaven.engine.table.Table;
+import io.deephaven.engine.updategraph.DynamicNode;
 import io.deephaven.engine.util.DelegatingScriptSession;
 import io.deephaven.engine.util.ExportedObjectType;
 import io.deephaven.engine.util.NoLanguageDeephavenSession;
 import io.deephaven.engine.util.ScriptSession;
 import io.deephaven.engine.util.VariableProvider;
-import io.deephaven.engine.updategraph.DynamicNode;
-import io.deephaven.figures.FigureWidgetTranslator;
+import io.deephaven.engine.util.jpy.JpyInit;
+import io.deephaven.extensions.barrage.util.GrpcUtil;
+import io.deephaven.grpc_api.console.figure.FigureWidgetTranslator;
 import io.deephaven.grpc_api.session.SessionCloseableObserver;
 import io.deephaven.grpc_api.session.SessionService;
 import io.deephaven.grpc_api.session.SessionState;
 import io.deephaven.grpc_api.session.SessionState.ExportBuilder;
 import io.deephaven.grpc_api.session.TicketRouter;
-import io.deephaven.extensions.barrage.util.GrpcUtil;
 import io.deephaven.internal.log.LoggerFactory;
 import io.deephaven.io.logger.LogBuffer;
 import io.deephaven.io.logger.LogBufferRecord;
@@ -34,8 +32,8 @@ import io.deephaven.lang.parse.CompletionParser;
 import io.deephaven.lang.parse.LspTools;
 import io.deephaven.lang.parse.ParsedDocument;
 import io.deephaven.lang.shared.lsp.CompletionCancelled;
+import io.deephaven.plot.FigureWidget;
 import io.deephaven.proto.backplane.grpc.Ticket;
-import io.deephaven.proto.backplane.script.grpc.*;
 import io.grpc.stub.StreamObserver;
 
 import javax.inject.Inject;
